complete restructure: two tsconfig files
