+++ .gitignore
@@ -1,4 +1,8 @@
 out
+media
 node_modules
 .vscode-test/
 *.vsix
+*~
+*.map
+package-lock.json
