+++ tsconfig.backend.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+	"module": "commonjs",
+	"target": "es2020",
+	"outDir": "out",
+	"rootDir": "src",
+	"lib": [
+	    "es2020",
+	    "dom"
+	],
+	"sourceMap": true,
+	"strict": false,   /* enable all strict type-checking options */
+	"esModuleInterop": true,
+	"skipLibCheck": true
+    },
+    "exclude": [
+	"src/webview/**/*.ts"
+    ]
+}
