--- tsconfig.json
@@ -1,24 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "commonjs",
-		"target": "es2020",
-		"outDir": "out",
-		"lib": [
-		    "es2020",
-		    "dom"
-		],
-		"sourceMap": true,
-		"rootDir": "src",
-		"strict": false,   /* enable all strict type-checking options */
-		"esModuleInterop": true,
-		"skipLibCheck": true
-		/* Additional Checks */
-		// "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
-		// "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
-		// "noUnusedParameters": true,  /* Report errors on unused parameters. */
-	},
-	"exclude": [
-		"node_modules",
-		".vscode-test"
-	]
-}
