+++ tsconfig.webview.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+	"module": "esnext",
+	"target": "es2020",
+	"moduleResolution": "bundler",
+	"rootDir": "src/webview",
+	"outDir": "media",
+	"lib": [
+	    "es2020",
+	    "dom"
+	],
+	"sourceMap": true,
+	"strict": false,   /* enable all strict type-checking options */
+	"esModuleInterop": true,
+	"skipLibCheck": true	
+    },
+    "include": [
+	"src/webview/**/*.ts"
+    ]
+}
